rng: widen to 32-bit so all 8 seed digits pick a distinct world
The RNG was a 16-bit LCG, so only 2^16 maps existed and the 8-hex-digit
seed was folded (s32 ^ s32>>16) into it — different seeds collided.

Replace it with xorshift32 (uint32 state), hand-coded in byte ops so no
32-bit shift/multiply helper is emitted (bank0_rng.o must stay self-
contained in BANK0 — banked item/monster code calls it, and reaching a
bank-1 helper while another bank is mapped crashes). The byte impl was
checked against canonical xorshift32 on the host (100k iters, 6 seeds).
rng_seed takes a uint32; rng_word returns the high 16 bits; the other
entry points are unchanged.

The title no longer folds — rng_seed(s32) takes the full 8 digits. The
saved RNG state g_save_rng widens 2->4 bytes; SAVE_VER 5->6 (old saves
reset). Existing verify_* tests search adaptively (no hard-coded map
coords), so all 14 still pass on the new maps — no re-baseline needed.

Proof: 0x00001234 / 0x12340000 / 0xFFFF1234 now hash to three distinct
maps; the first two (which folded equal before) differ. Save roundtrip
carries the full 32-bit state (upper half nonzero).

Note: a random new game still seeds from 16-bit entropy (DIV<<8|frames);
the seed editor is what unlocks the full 2^32.

+++ src/rng.h
@@ -3,8 +3,8 @@
 
 #include <stdint.h>
 
-void     rng_seed(uint16_t seed);
-uint16_t rng_state(void);        /* for save.c */
+void     rng_seed(uint32_t seed);
+uint32_t rng_state(void);        /* for save.c */
 uint16_t rng_word(void);
 uint8_t  rng_byte(void);
 uint8_t  rng_range(uint8_t n);   /* 0..n-1 */
